--- js/models/model.js
@@ -1,14 +1,0 @@
-import { fetchData } from '../components/service.js';
-
-const model = {
-  init() {
-    this.data = {};
-    this.loadData();
-  },
-  loadData() {
-    this.data = await fetchData();
-    return this.data;
-  },
-};
-
-export { model };
